+++ src/ru/VladTheMountain/oclide/emulator/brain/Emulator.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2021 Vladislav Gorskii.
+ * Copyright 2021 Vladi.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,28 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package ru.VladTheMountain.oclide.util;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import javax.swing.JTextArea;
+package ru.VladTheMountain.oclide.emulator.brain;
 
 /**
- * Util class for redirecting {@link System.out} to {@code outputArea}
+ * TODO: Add configurable emulation from {@link EmulatorFrame}
  *
- * @author VladTheMountain
+ * @author Vladi
  */
-public class ConsoleOutputStream extends OutputStream {
-
-    private JTextArea target;
-
-    public ConsoleOutputStream(JTextArea area) {
-        target = area;
-    }
-
-    @Override
-    public void write(int b) throws IOException {
-        target.append(String.valueOf((char) b));
-    }
+public class Emulator {
 
 }
