+++ src/ru/VladTheMountain/oclide/resources/emulator/brain/OcelotBrain.properties
